Introduce hole filling

+++ libs/tex/TextureView.cpp
@@ -4,6 +4,7 @@ TextureView::TextureView(std::size_t id, mve::CameraInfo const & camera,
     mve::ByteImage::Ptr image)
     : id(id), image(image) {
 
+    //mve::image::gamma_correct(this->image, 2.2f);
     width = image->width();
     height = image->height();
 
@@ -189,12 +190,12 @@ TextureView::get_face_info(math::Vec3f const & v1, math::Vec3f const & v2,
 
     if (settings.data_term == GMI) {
         if (num_samples > 0) {
-            gmi = (gmi / num_samples) * static_cast<double>(area);
+            gmi = (gmi / num_samples) * area;
         } else {
             double gmv1 = static_cast<double>(gradient_magnitude->linear_at(p1[0], p1[1], 0)) / 255.0;
             double gmv2 = static_cast<double>(gradient_magnitude->linear_at(p2[0], p2[1], 0)) / 255.0;
             double gmv3 = static_cast<double>(gradient_magnitude->linear_at(p3[0], p3[1], 0)) / 255.0;
-            gmi = ((gmv1 + gmv2 + gmv3) / 3.0) * static_cast<double>(area);
+            gmi = ((gmv1 + gmv2 + gmv3) / 3.0) * area;
         }
     }
